requirements.txt
pytest
requests
requests-mock
pyyaml
openpyxl
pandas
pydantic[dotenv]